+++ cirruslib/__init__.py
@@ -1,4 +1,4 @@
 from .catalog import Catalog, Catalogs
 from .logging import get_task_logger
-from .statedb import StateDB
+from .statedb import StateDB, STATES
 from .version import __version__