+++ tests/test_statedb.py
@@ -12,13 +12,18 @@ from datetime import datetime
 from decimal import Decimal
 from cirruslib.statedb import StateDB, STATES
 
-
 ## fixtures
 testpath = os.path.dirname(__file__)
 table_name = 'cirrus-test-state'
-
+test_dbitem = {
+    'collections_workflow': 'col1_wf1',
+    'itemids': 'item1/item2',
+    'state_updated': f"QUEUED_{datetime.now()}",
+    'created': datetime.now(),
+    'updated': datetime.now()
+}
 test_item = {
-    "id": "input-collection/workflow-test/item",
+    "id": "col1/workflow-wf1/item1/item2",
     "process": {
         "output_options": {
             "collections": {
@@ -28,41 +33,28 @@ test_item = {
     }
 }
 
-test_dbitem = {
-    'input_collections': 'input-collection',
-    'output_collections': 'output-collection',
-    'id': 'test/item',
-    'current_state': f"QUEUED_{datetime.now()}",
-    'created_at': datetime.now()
-}
-
-
-def setUpModule():
+@mock_dynamodb2
+def setup_table():
+    boto3.setup_default_session()
     client = boto3.resource('dynamodb')
     with open(os.path.join(testpath, 'statedb_schema.json')) as f:
         schema = json.loads(f.read())
     table = client.create_table(**schema)
     table.meta.client.get_waiter('table_exists').wait(TableName=table_name)
-
-
-def tearDownModule():
-    client = boto3.resource('dynamodb')
-    table = client.Table(table_name)
-    table.delete()
-    table.wait_until_not_exists()
+    return StateDB(table_name) 
 
 
 class TestClassMethods(unittest.TestCase):
 
     testkey = {
-        'input_collections': 'input-collection',
-        'id': 'test/item'
+        'collections_workflow': 'col1_wf1',
+        'itemids': 'item1/item2'
     }
 
     def test_catid_to_key(self):
         key = StateDB.catid_to_key(test_item['id'])
-        assert(key['input_collections'] == "input-collection")
-        assert(key['id'] == 'test/item')
+        assert(key['collections_workflow'] == "col1_wf1")
+        assert(key['itemids'] == 'item1/item2')
 
     def test_key_to_catid(self):
         catid = StateDB.key_to_catid(self.testkey)
@@ -75,7 +67,7 @@ class TestClassMethods(unittest.TestCase):
     def test_dbitem_to_item(self):
         item = StateDB.dbitem_to_item(test_dbitem)
         assert(item['catid'] == test_item['id'])
-        assert(item['workflow'] == 'test')
+        assert(item['workflow'] == 'wf1')
         assert(item['state'] == 'QUEUED')
 
     def test_since_to_timedelta(self):
@@ -87,37 +79,58 @@ class TestClassMethods(unittest.TestCase):
         assert(td.seconds == 600)
 
 
-# TODO - figure out why mocking still sends queries to AWS
-#@mock_dynamodb2
 class TestDbItems(unittest.TestCase):
+
+    nitems = 1000
 
     @classmethod
     def setUpClass(cls):
-        cls.statedb = StateDB(table_name)
-        for i in range(10):
+        cls.mock = mock_dynamodb2()
+        cls.mock.start()
+        cls.statedb = setup_table()
+        for i in range(cls.nitems):
             newitem = deepcopy(test_item)
-            newitem['id'] = newitem['id'] + str(i)
-            cls.statedb.create_item(newitem)
+            cls.statedb.set_processing(newitem['id'] + str(i), execution='arn::test')
+        cls.statedb.set_processing(test_item['id'] + '_processing', execution='arn::test')
+        cls.statedb.set_completed(test_item['id'] + '_completed', outputs=['item1', 'item2'])
+        cls.statedb.set_failed(test_item['id'] + '_failed', 'failed')
+        cls.statedb.set_invalid(test_item['id'] + '_invalid', 'invalid')
+
 
     @classmethod
     def tearDownClass(cls):
-        for i in range(10):
+        for i in range(cls.nitems):
             cls.statedb.delete_item(test_item['id'] + str(i))
+        for s in STATES:
+            cls.statedb.delete_item(test_item['id'] + f"_{s.lower()}")
+        cls.statedb.delete()
+        cls.mock.stop()
 
-    def test_create_item(self):
-        resp = self.statedb.create_item(test_item)
+    def test_set_processing(self):
+        resp = self.statedb.set_processing(test_item['id'], execution='arn::test1')
         assert(resp['ResponseMetadata']['HTTPStatusCode'] == 200)
         dbitem = self.statedb.get_dbitem(test_item['id'])
-        assert(dbitem['id'] == test_dbitem['id'])
+        assert(StateDB.key_to_catid(dbitem) == test_item['id'])
+        assert(dbitem['executions'] == ['arn::test1'])
+
+        # check that processing adds new execution to list
+        resp = self.statedb.set_processing(test_item['id'], execution='arn::test2')
+        dbitem = self.statedb.get_dbitem(test_item['id'])
+        assert(len(dbitem['executions']) == 2)
+        assert(dbitem['executions'][-1] == 'arn::test2')
         self.statedb.delete_item(test_item['id'])
-        #with self.assertRaises(Exception):
-        #    dbitem = self.statedb.get_dbitem(test_item['id'])
+        dbitem = self.statedb.get_dbitem(test_item['id'])
+        assert(dbitem is None)
 
     def test_get_dbitem(self):
-        item = self.statedb.get_dbitem(test_item['id'] + '0')
-        assert(item['id'] == test_dbitem['id'] + '0')
-        assert(item['input_collections'] == test_dbitem['input_collections'])
-        assert(item['current_state'].startswith('QUEUED'))
+        dbitem = self.statedb.get_dbitem(test_item['id'] + '0')
+        assert(dbitem['itemids'] == test_dbitem['itemids'] + '0')
+        assert(dbitem['collections_workflow'] == test_dbitem['collections_workflow'])
+        assert(dbitem['state_updated'].startswith('PROCESSING'))
+
+    def test_get_dbitem_noitem(self):
+        dbitem = self.statedb.get_dbitem(test_item['id'])
+        assert(dbitem is None)
 
     def test_get_dbitems(self):
         ids = [test_item['id'] + str(i) for i in range(10)]
@@ -126,31 +139,25 @@ class TestDbItems(unittest.TestCase):
         for dbitem in dbitems:
             assert(self.statedb.key_to_catid(dbitem) in ids)
 
+    def test_get_dbitems_noitems(self):
+        #with self.assertRaises(Exception):
+        dbitems = self.statedb.get_dbitems([test_item['id']])
+        assert(len(dbitems) == 0)
 
-class TestStates(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        cls.statedb = StateDB(table_name)
-        cls.statedb.create_item(test_item, state='QUEUED')
-        for i, s in enumerate(STATES):
-            newitem = deepcopy(test_item)
-            newitem['id'] = newitem['id'] + str(i)
-            cls.statedb.create_item(newitem, state=s)
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.statedb.delete_item(test_item['id'])
-        for i in range(len(STATES)):
-            cls.statedb.delete_item(test_item['id'] + str(i))        
+    def test_get_items(self):
+        items = self.statedb.get_items(test_dbitem['collections_workflow'], state='PROCESSING', since='1h')
+        assert(len(items) == self.nitems + 1)
+        items = self.statedb.get_items(test_dbitem['collections_workflow'], state='PROCESSING', since='1h', limit=1)
+        assert(len(items) == 1)
 
     def test_get_state(self):
-        for i, s in enumerate(STATES):
-            state = self.statedb.get_state(test_item['id'] + str(i))
+        for s in STATES:
+            state = self.statedb.get_state(test_item['id'] + f"_{s.lower()}")
             assert(state == s)
+        state = self.statedb.get_state(test_item['id'] + 'nosuchitem')
 
     def test_get_states(self):
-        ids = [test_item['id'] + str(i) for i in range(len(STATES))]
+        ids = [test_item['id'] + f"_{s.lower()}" for s in STATES]
         states = self.statedb.get_states(ids)
         assert(len(ids) == len(states))
         for i, id in enumerate(ids):
@@ -160,43 +167,46 @@ class TestStates(unittest.TestCase):
         resp = self.statedb.set_processing(test_item['id'], execution='testarn')
         assert(resp['ResponseMetadata']['HTTPStatusCode'] == 200)
         dbitem = self.statedb.get_dbitem(test_item['id'])
-        assert(dbitem['current_state'].startswith('PROCESSING'))
-        assert(dbitem['execution'] == 'testarn')
+        assert(dbitem['state_updated'].startswith('PROCESSING'))
+        assert(dbitem['executions'] == ['testarn'])
 
-    def test_set_complete(self):
-        resp = self.statedb.set_completed(test_item['id'], urls=['output-url'])
+    def test_set_completed(self):
+        resp = self.statedb.set_completed(test_item['id'], outputs=['output-item'])
         assert(resp['ResponseMetadata']['HTTPStatusCode'] == 200)
         dbitem = self.statedb.get_dbitem(test_item['id'])
-        assert(dbitem['current_state'].startswith('COMPLETED'))
-        assert(dbitem['output_urls'][0] == 'output-url')
+        assert(dbitem['state_updated'].startswith('COMPLETED'))
+        assert(dbitem['outputs'][0] == 'output-item')
 
     def test_set_failed(self):
         resp = self.statedb.set_failed(test_item['id'], msg='test failure')
         assert(resp['ResponseMetadata']['HTTPStatusCode'] == 200)
         dbitem = self.statedb.get_dbitem(test_item['id'])
-        assert(dbitem['current_state'].startswith('FAILED'))
+        assert(dbitem['state_updated'].startswith('FAILED'))
+        assert(dbitem['last_error'] == 'test failure')
 
     def test_set_invalid(self):
         resp = self.statedb.set_invalid(test_item['id'], msg='test failure')
         assert(resp['ResponseMetadata']['HTTPStatusCode'] == 200)
         dbitem = self.statedb.get_dbitem(test_item['id'])
-        assert(dbitem['current_state'].startswith('INVALID'))
+        assert(dbitem['state_updated'].startswith('INVALID'))
+        assert(dbitem['last_error'] == 'test failure')
 
-# TODO - figure out why mocking still sends queries to AWS
-#@mock_dynamodb2
-class TestItems(unittest.TestCase):
+    def test_get_counts(self):
+        count = self.statedb.get_counts(test_dbitem['collections_workflow'])
+        assert(count == self.nitems + 4)
+        for s in STATES:
+            count = self.statedb.get_counts(test_dbitem['collections_workflow'], state=s)
+            if s == 'PROCESSING':
+                assert(count == self.nitems + 1)
+            else:
+                assert(count == 1)
+        count = self.statedb.get_counts(test_dbitem['collections_workflow'], since='1h')
 
-    @classmethod
-    def setUpClass(cls):
-        cls.statedb = StateDB(table_name)
-        cls.statedb.create_item(test_item, state='QUEUED')
-        for i, s in enumerate(STATES):
-            newitem = deepcopy(test_item)
-            newitem['id'] = newitem['id'] + str(i)
-            cls.statedb.create_item(newitem, state=s)
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.statedb.delete_item(test_item['id'])
-        for i in range(len(STATES)):
-            cls.statedb.delete_item(test_item['id'] + str(i))     
+        
+    def _test_get_counts_paging(self):
+        for i in range(5000):
+            self.statedb.set_processing(test_item['id'] + f"_{i}", execution='arn::test')
+        count = self.statedb.get_counts(test_dbitem['collections_workflow'])
+        assert(count == 1004)
+        for i in range(5000):
+            self.statedb.delete_item(test_item['id'] + f"_{i}")