+++ tests/test_transfer.py
@@ -15,7 +15,8 @@ testbucket = 'testbucket'
 class Test(unittest.TestCase):
 
     def setUp(self):
-        client = s3()
+        session = boto3.session.Session(region_name='us-east-1')
+        client = s3(session)
         client.s3.create_bucket(Bucket=testbucket)
         client.s3.put_object(Body='test', Bucket=testbucket, Key=os.path.basename(__file__))
         os.makedirs(testpath, exist_ok=True)
@@ -46,7 +47,7 @@ class Test(unittest.TestCase):
         buckets = session.s3.list_buckets()
         assert(buckets['Buckets'][0]['Name'] == testbucket)
 
-    def test_download_from_http(self):
+    def _test_download_from_http(self):
         url = 'https://raw.githubusercontent.com/cirrus-geo/cirrus/master/README.md'
         fname = transfer.download_from_http(url, path=testpath)
         assert(os.path.exists(fname))